getting horrible shadow acne; staging for now

+++ function.h
@@ -2,7 +2,7 @@
 #include "rng.h"
 #include <cstdint>
 
-glm::vec3 trace_ray(const Ray &ray, const Scene &scene, RNG &rng,
-                    int16_t bounces);
+glm::dvec3 trace_ray(const Ray &ray, const Scene &scene, RNG &rng,
+                     int16_t bounces);
 
 void render(const Scene &scene, const Camera &camera, Film &film);
