+++ util.h
@@ -1,4 +1,5 @@
 #pragma once
 #include "data.h"
-bool ray_triangle_intersect(const Ray &ray, const glm::vec3 &v0,
-                            const glm::vec3 &v1, const glm::vec3 &v2, float &t);
+bool ray_triangle_intersect(const Ray &ray, const glm::dvec3 &v0,
+                            const glm::dvec3 &v1, const glm::dvec3 &v2,
+                            double &t);
